Move history class to own module (net.py already very long).
