+++ inferno/__init__.py
@@ -1,3 +1,15 @@
-from .net import History, NeuralNet, NeuralNetClassifier, NeuralNetRegressor
+from .history import History
+from .net import NeuralNet
+from .net import NeuralNetClassifier
+from .net import NeuralNetRegressor
 
 from . import callbacks
+
+
+__all__ = [
+    'History',
+    'NeuralNet',
+    'NeuralNetClassifier',
+    'NeuralNetRegressor',
+    'callbacks',
+]
